requirements.txt
uvicorn[standard]
fastapi~=0.115.12
alembic~=1.15.2
psycopg2-binary
sqlmodel~=0.0.24
SQLAlchemy~=2.0.40
asyncpg
requests
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
bcrypt<4.0